requirements.txt
# Core training stack — all versions pinned
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# HuggingFace ecosystem
transformers==4.44.2
peft==0.12.0
trl==0.10.1
datasets==2.21.0
accelerate==0.33.0
tokenizers==0.19.1
huggingface-hub==0.24.6
safetensors==0.4.4

# Quantization
bitsandbytes==0.43.3

# Structured output / constrained decoding
outlines==0.0.46
instructor==1.3.3

# Experiment tracking (all optional at runtime)
wandb==0.17.8
mlflow==2.15.1

# Data / storage
boto3==1.35.7
s3fs==2024.6.1
pyarrow==17.0.0

# Evaluation helpers
jsonschema==4.23.0
nltk==3.9.1
rouge-score==0.1.2
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2

# Config / CLI
pyyaml==6.0.2
omegaconf==2.3.0
click==8.1.7
rich==13.7.1

# Containerization / reproducibility
python-dotenv==1.0.1

# Dev / testing
pytest==8.3.2
pytest-cov==5.0.0
